[REFACTORING] : chatList MVVM 완성 - firebase코드 viewmodel로 모두 이동
