+++ app/src/main/java/com/erionna/eternalreturninfo/presentation/activity/board/BoardPostActivity.kt
@@ -13,9 +13,9 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import coil.api.load
 import com.erionna.eternalreturninfo.R
 import com.erionna.eternalreturninfo.databinding.BoardPostActivityBinding
-import com.erionna.eternalreturninfo.model.BoardModel
-import com.erionna.eternalreturninfo.model.CommentModel
-import com.erionna.eternalreturninfo.model.ERModel
+import com.erionna.eternalreturninfo.data.model.BoardModel
+import com.erionna.eternalreturninfo.data.model.CommentModel
+import com.erionna.eternalreturninfo.data.model.ERModel
 import com.erionna.eternalreturninfo.retrofit.BoardSingletone
 import com.erionna.eternalreturninfo.retrofit.FBRef
 import com.erionna.eternalreturninfo.presentation.activity.chat.ChatActivity
