+++ app/src/main/java/com/erionna/eternalreturninfo/presentation/adapter/chat/ChatAdapter2.kt
@@ -10,7 +10,7 @@ import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.erionna.eternalreturninfo.databinding.ChatItemReceiverBinding
 import com.erionna.eternalreturninfo.databinding.ChatItemSenderBinding
-import com.erionna.eternalreturninfo.model.Message
+import com.erionna.eternalreturninfo.data.model.Message
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.database.DataSnapshot
 import com.google.firebase.database.DatabaseError
