+++ app/src/main/java/com/erionna/eternalreturninfo/presentation/fragment/board/BaseFragment.kt
@@ -13,7 +13,7 @@ import androidx.activity.result.contract.ActivityResultContracts
 import androidx.fragment.app.activityViewModels
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.erionna.eternalreturninfo.databinding.BoardRvFragmentBinding
-import com.erionna.eternalreturninfo.model.BoardModel
+import com.erionna.eternalreturninfo.data.model.BoardModel
 import com.erionna.eternalreturninfo.retrofit.FBRef
 import com.erionna.eternalreturninfo.presentation.activity.board.BoardDeletedActivity
 import com.erionna.eternalreturninfo.presentation.activity.board.BoardPostActivity
