+++ app/src/main/java/com/erionna/eternalreturninfo/presentation/fragment/board/BoardFragment.kt
@@ -12,7 +12,7 @@ import androidx.activity.result.contract.ActivityResultContracts
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import com.erionna.eternalreturninfo.databinding.BoardFragmentBinding
-import com.erionna.eternalreturninfo.model.BoardModel
+import com.erionna.eternalreturninfo.data.model.BoardModel
 import com.erionna.eternalreturninfo.retrofit.FBRef
 import com.erionna.eternalreturninfo.presentation.activity.board.BoardAddActivity
 import com.erionna.eternalreturninfo.presentation.activity.board.BoardSearchActivity
