+++ app/src/main/java/com/erionna/eternalreturninfo/presentation/viewmodel/ChatListViewModel.kt
@@ -1,19 +1,42 @@
 package com.erionna.eternalreturninfo.presentation.viewmodel
 
+import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.erionna.eternalreturninfo.data.repository.ChatListRepositoryImpl
-import com.erionna.eternalreturninfo.model.ERModel
+import coil.api.load
+import com.erionna.eternalreturninfo.data.model.ERModel
+import com.erionna.eternalreturninfo.data.model.Message
+import com.erionna.eternalreturninfo.presentation.fragment.chat.ChatListFragment
 import com.erionna.eternalreturninfo.presentation.repository.ChatListRepository
-import java.util.concurrent.atomic.AtomicLong
+import com.google.firebase.auth.ktx.auth
+import com.google.firebase.database.DataSnapshot
+import com.google.firebase.database.DatabaseError
+import com.google.firebase.database.DatabaseReference
+import com.google.firebase.database.ValueEventListener
+import com.google.firebase.database.ktx.database
+import com.google.firebase.ktx.Firebase
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.launch
 
 class ChatListViewModel(
     val repository: ChatListRepository
 ) : ViewModel() {
     private val _list: MutableLiveData<List<ERModel>> = MutableLiveData()
     val list: LiveData<List<ERModel>> get() = _list
+
+    private val _profilePictureUrl: MutableLiveData<String> = MutableLiveData()
+    val profilePictureUrl: LiveData<String> get() = _profilePictureUrl
+
+    private val _profileName: MutableLiveData<String?> = MutableLiveData()
+    val profileName: LiveData<String?> get() = _profileName
+
+
+    private lateinit var database: DatabaseReference
+    private var auth = Firebase.auth
+    private lateinit var refEventListener: ValueEventListener
+    private lateinit var refDb: DatabaseReference
 
     fun addUser(
         item: ERModel?
@@ -33,18 +56,99 @@ class ChatListViewModel(
         _list.value = repository.modifyItemForChatList(item)
     }
 
-}
+    fun addChatList() {
+        database = Firebase.database.reference
+        refDb = database.child("user")
+        refEventListener = object : ValueEventListener {
 
-class ChatListViewModelFactory : ViewModelProvider.Factory {
-    private val repository: ChatListRepository = ChatListRepositoryImpl(
-        AtomicLong(1L)
-    )
+            override fun onDataChange(snapshot: DataSnapshot) {
+                // 리스트 초기화 for 회원가입시 리스트 중복추가 문제 해결
+                clearList()
+                var senderRoom = ""
+                var receiverRoom = ""
 
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(ChatListViewModel::class.java)) {
-            return ChatListViewModel(repository) as T
-        } else {
-            throw IllegalArgumentException("not found correct viewModel")
+                for (child in snapshot.children) {
+                    val currentUser = child.getValue(ERModel::class.java)
+
+                    senderRoom = currentUser?.uid + auth.currentUser?.uid
+                    var message = Message()
+
+                    database.child("chats").child(senderRoom).child("messages")
+                        .addValueEventListener(object : ValueEventListener {
+                            override fun onDataChange(snapshot: DataSnapshot) {
+                                for (child in snapshot.children) {
+                                    message = child.getValue(Message::class.java)!!
+                                }
+                                if (message.whereRU == true) {
+                                    addUser(
+                                        currentUser?.copy(
+                                            msg = message.message,
+                                            time = message.time,
+                                            readOrNot = message.readOrNot
+                                        )
+                                    )
+                                }
+                            }
+
+                            override fun onCancelled(error: DatabaseError) {
+                                Log.e("choco5733 in chatList", error.message)
+                            }
+
+                        })
+
+                    database.child("chats").child(senderRoom).child("messages")
+                        .addValueEventListener(object : ValueEventListener {
+                            override fun onDataChange(snapshot: DataSnapshot) {
+
+                                for (child in snapshot.children) {
+                                    message = child.getValue(Message::class.java)!!
+                                }
+                                Log.d("choco5733 in msg", "$message")
+
+
+                                modifyItemForChatList(
+                                    currentUser?.copy(
+                                        msg = message.message,
+                                        time = message.time,
+                                        readOrNot = message.readOrNot
+                                    )
+                                )
+                            }
+
+                            override fun onCancelled(error: DatabaseError) {
+                                Log.e("choco5733", error.message)
+                            }
+                        })
+
+                    if (auth.uid == currentUser?.uid) {
+                        _profileName.value = currentUser?.name
+                    }
+
+                }
+            }
+
+            override fun onCancelled(error: DatabaseError) {
+                Log.e("choco5733", error.message)
+            }
+
+        }
+        refDb.addValueEventListener(refEventListener)
+    }
+
+    fun disconnectFirebase() {
+        refDb.removeEventListener(refEventListener)
+    }
+
+    fun loadProfilePicture()  {
+        database = Firebase.database.reference
+        database.child("user").get().addOnSuccessListener {
+            for(child in it.children) {
+                val currentUser = child.getValue(ERModel::class.java)
+                if (currentUser?.uid == auth.uid) {
+                    _profilePictureUrl.value = currentUser?.profilePicture.toString()
+                }
+            }
         }
     }
 }
+
