+++ app/src/main/java/com/erionna/eternalreturninfo/presentation/viewmodel/ChatViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.erionna.eternalreturninfo.data.repository.ChatRepositoryImpl
-import com.erionna.eternalreturninfo.model.Message
+import com.erionna.eternalreturninfo.data.model.Message
 import com.erionna.eternalreturninfo.presentation.repository.ChatRepository
 
 class ChatViewModel(
@@ -27,17 +27,5 @@ class ChatViewModel(
 
     fun getListSize(): Int {
         return repository.getListSize()
-    }
-}
-
-class ChatViewModelFactory : ViewModelProvider.Factory {
-    private val repository: ChatRepository = ChatRepositoryImpl()
-
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(ChatViewModel::class.java)) {
-            return ChatViewModel(repository) as T
-        } else {
-            throw IllegalArgumentException("not found correct viewModel")
-        }
     }
 }
